:truck: chore: Move DoctorActivity to ui folder
#60
