+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:theme="@style/Theme.DermatologyApp"
         tools:targetApi="31">
         <activity
-            android:name=".DoctorActivity"
+            android:name=".ui.DoctorActivity"
             android:exported="false" />
         <activity
             android:name=".ui.NotificationActivity"
