+++ app/src/main/java/br/com/ilstudio/dermatologyapp/ui/DoctorActivity.kt
@@ -1,4 +1,4 @@
-package br.com.ilstudio.dermatologyapp
+package br.com.ilstudio.dermatologyapp.ui
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
